Add tree staves for bowmaking
The name is questionable, but I can't find anything other than bow stave for this particular cut, and I don't want players to think that's all they can do with it. I'm assuming these are roughly twice the size of a long stick, but they're more wedge shaped than round so aren't as good as a melee weapon, but do a bit more damage by weight alone. Each stave could then be split again, giving something reasonably similar to a long stick and a couple of 2x4s. That way the player isn't left with a bunch of useless bow blanks after going to all the effort of cutting down a tree and splitting it. These will be mainly for longbows but might be useful for some crossbows and anything else that requires a long, thick, straight-grain piece of wood (giggity). There's an argument to be made for requiring them for polearms handles but that's not my jam.