Add HAMMER_HEAVY
New quality for things like the sledgehammer. Criteria is that it has a long handle and/or very heavy head, as well as a surface suitable for hammering. Expected usage is driving wedges to split logs, or things like pounding stakes into the ground. Current tools in mind are the sledgehammer, wood axe (if turned to hit with the flat side), halligan bar, and the war hammer. Could probably give level 1 to makeshift heavy hammers like the halligan bar and warhammer, then level 2 to dedicated ones like the sledgehammer or a maul, but I only actually have one thing in mind for this to work with so I'm not going to bother for now.