Allow splitting tree trunks into staves
Plus some extra goodies.  Tends to be a fairly wasteful process, so there's some offcuts. Follows conservation of mass by volume as best I can, but most other tree trunk activities seem to be fairly wonky, so this is probably good enough until there's solid numbers to base things on.

Splitting requires wedges, but you can use wooden wedges in a pinch. My assumption is that there would be plenty of branches and sticks around after a tree is felled, so the player uses the cutting tool to cut some wedges and uses those to split the log.